Refactor `Source.py` into `Import.py` and `Segment.py`
The `Source.py` file contained the `Source`, `Import`, `ImportList`, `Segment`, and `SegmentList` classes. This made the file large and difficult to maintain.

I have refactored the `Source.py` file by moving the `Import` and `ImportList` classes to a new `Import.py` file, and the `Segment` and `SegmentList` classes to a new `Segment.py` file. The `Source.py` file now only contains the `Source` class.

I also updated the `__init__.py` file to import the classes from their new locations and updated the test files to reflect the new file structure.

All quality checks (`pydocstyle`, `pyright`, `pytest`) pass after this refactoring.
