+++ PermutiveAPI/Segment.py
@@ -1,0 +1,312 @@
+"""Segment management for the Permutive API."""
+
+import logging
+from typing import Dict, List, Optional
+from dataclasses import dataclass, field
+from datetime import datetime
+
+from PermutiveAPI.Utils import RequestHelper, JSONSerializable
+
+
+_API_VERSION = "v1"
+_API_ENDPOINT = f'https://api.permutive.app/audience-api/{_API_VERSION}/imports'
+_API_PAYLOAD = ['name', 'code', 'description', 'cpm', 'categories']
+
+
+@dataclass
+class Segment(JSONSerializable):
+    """
+    Represent a segment in the Permutive API.
+
+    Attributes:
+        code (str): The code of the segment.
+        name (str): The name of the segment.
+        import_id (str): The import ID of the segment.
+        id (Optional[str]): The ID of the segment.
+        description (Optional[str]): The description of the segment.
+        cpm (Optional[float]): The cost per mille of the segment.
+        categories (Optional[List[str]]): Categories associated with the segment.
+        updated_at (Optional[datetime]): When the segment was last updated.
+    """
+
+    code: str
+    name: str
+    import_id: str
+    id: Optional[str] = None
+    description: Optional[str] = None
+    cpm: Optional[float] = 0.0
+    categories: Optional[List[str]] = None
+    updated_at: Optional[datetime] = field(default_factory=datetime.now)
+
+    def create(self,
+               api_key: str):
+        """Create a new segment using the provided private key.
+
+        Args:
+            api_key (str): The private key used for authentication.
+
+        Returns:
+            None
+
+        Raises:
+            ValueError: If the segment creation fails.
+        """
+        logging.debug(
+            f"SegmentAPI::create_segment::{self.import_id}::{self.name}")
+        url = f"{_API_ENDPOINT}/{self.import_id}/segments"
+        response = RequestHelper.post_static(api_key=api_key,
+                                             url=url,
+                                             data=RequestHelper.to_payload_static(dataclass_obj=self,
+                                                                                  api_payload=_API_PAYLOAD))
+        if not response:
+            raise ValueError('Unable to create_segment')
+
+        new_segment = Segment.from_json(response.json())
+        if isinstance(new_segment, Segment):
+            self.__dict__.update(new_segment.__dict__)
+
+
+    def update(self,
+               api_key: str):
+        """Update the segment using the provided private key.
+
+        Args:
+            api_key (str): The private key used for authentication.
+
+        Returns:
+            None
+
+        Raises:
+            ValueError: If the segment update fails.
+        """
+        logging.debug(
+            f"SegmentAPI::update_segment::{self.import_id}::{self.name}")
+        url = f"{_API_ENDPOINT}/{self.import_id}/segments/{self.id}"
+        response = RequestHelper.patch_static(api_key=api_key,
+                                              url=url,
+                                              data=RequestHelper.to_payload_static(dataclass_obj=self,
+                                                                                   api_payload=_API_PAYLOAD))
+        if not response:
+            raise ValueError('Unable to update_segment')
+
+        updated_segment = Segment.from_json(response.json())
+        if isinstance(updated_segment, Segment):
+            self.__dict__.update(updated_segment.__dict__)
+
+    def delete(self,
+               api_key: str) -> bool:
+        """Delete a segment using the provided private key.
+
+        Args:
+            api_key (str): The private key used for authentication.
+
+        Returns:
+            bool: ``True`` if the segment was successfully deleted (status code
+                204), otherwise ``False``.
+        """
+        logging.debug(
+            f"SegmentAPI::delete_segment::{self.import_id:}::{self.id}")
+        url = f"{_API_ENDPOINT}/{self.import_id}/segments/{self.id}"
+        response = RequestHelper.delete_static(api_key=api_key,
+                                               url=url)
+        if response is None:
+            raise ValueError("Response is None")
+        return response.status_code == 204
+
+    @staticmethod
+    def get_by_code(import_id: str,
+                    segment_code: str,
+                    api_key: str) -> 'Segment':
+        """Retrieve a segment by its code.
+
+        Args:
+            import_id (str): The ID of the import.
+            segment_code (str): The code of the segment to retrieve.
+            api_key (str): The private key for authentication.
+
+        Returns:
+            Segment: The segment object retrieved by the given code.
+
+        Raises:
+            ValueError: If the segment cannot be retrieved.
+        """
+        logging.debug(
+            f"SegmentAPI::get_segment_by_code::{import_id}::{segment_code}")
+        url = f"{_API_ENDPOINT}/{import_id}/segments/code/{segment_code}"
+        response = RequestHelper.get_static(url=url, api_key=api_key
+                                            )
+        if not response:
+            raise ValueError('Unable to get_segment')
+        segment = Segment.from_json(response.json())
+        if isinstance(segment, Segment):
+            return segment
+        else:
+            raise ValueError('Segment not found or invalid response format')
+
+    @staticmethod
+    def get_by_id(import_id: str,
+                  segment_id: str,
+                  api_key: str) -> 'Segment':
+        """Retrieve a segment by its ID.
+
+        Args:
+            import_id (str): The ID of the import.
+            segment_id (str): The ID of the segment to retrieve.
+            api_key (str): The private key for authentication.
+
+        Returns:
+            Segment: The segment object retrieved by the given ID.
+
+        Raises:
+            ValueError: If the segment cannot be retrieved.
+        """
+        logging.debug(
+            f"SegmentAPI::get_segment_by_id::{import_id}::{segment_id}")
+        url = f"{_API_ENDPOINT}/{import_id}/segments/{segment_id}"
+        response = RequestHelper.get_static(url=url,
+                                            api_key=api_key)
+        if not response:
+            raise ValueError('Unable to get_by_id')
+        segment = Segment.from_json(response.json())
+        if isinstance(segment, Segment):
+            return segment
+        else:
+            raise ValueError('Segment not found or invalid response format')
+
+    @staticmethod
+    def list(import_id: str,
+             api_key: str) -> List['Segment']:
+        """Retrieve a list of segments for a given import ID.
+
+        Args:
+            import_id (str): The ID of the import to retrieve segments for.
+            api_key (str): The private key used for authentication.
+
+        Returns:
+            List[Segment]: A list of Segment objects retrieved from the API.
+
+        Raises:
+            requests.exceptions.RequestException: If an error occurs while making
+                the API request.
+        """
+        logging.debug(f"SegmentAPI::list")
+
+        base_url = f"{_API_ENDPOINT}/{import_id}/segments"
+        all_segments = []
+        next_token = None
+
+        while True:
+            # Construct the URL with the pagination token
+            url = f"{base_url}?pagination_token={next_token}" if next_token else base_url
+            response = RequestHelper.get_static(
+                api_key=api_key, url=url)
+            if response is None:
+                raise ValueError("Response is None")
+            data = response.json()
+
+            # Extract elements and add them to the list
+            all_segments.extend([Segment.from_json(element)
+                                for element in data.get('elements', [])])
+
+            # Check for next_token in the pagination metadata
+            next_token = data.get('pagination', {}).get('next_token')
+
+            if not next_token:
+                break  # Stop when there are no more pages
+
+        return all_segments
+
+
+class SegmentList(List[Segment],
+                  JSONSerializable):
+    """
+    Custom list that holds Segment objects and provides caching and serialization.
+
+    Attributes:
+        _id_dictionary_cache (Dict[str, Segment]): Cache mapping segment IDs to objects.
+        _name_dictionary_cache (Dict[str, Segment]): Cache mapping segment names to objects.
+        _code_dictionary_cache (Dict[str, Segment]): Cache mapping segment codes to objects.
+    """
+
+    def __init__(self,
+                 items_list: Optional[List[Segment]] = None):
+        """Initialize the SegmentList with an optional list of Segment objects.
+
+        Args:
+            items_list (Optional[List[Segment]]): Segment objects to initialize
+                with.
+
+        Returns:
+            None
+        """
+        super().__init__(items_list if items_list is not None else [])
+        self.rebuild_cache()
+
+    def rebuild_cache(self):
+        """Rebuild all caches based on the current state of the list.
+
+        Returns:
+            None
+        """
+        self._id_dictionary_cache = {
+            segment.id: segment for segment in self if segment.id}
+        self._name_dictionary_cache = {
+            segment.name: segment for segment in self if segment.name}
+        self._code_dictionary_cache = {
+            segment.code: segment for segment in self if segment.code}
+
+    @property
+    def id_dictionary(self) -> Dict[str, Segment]:
+        """Return a dictionary of segments indexed by their IDs.
+
+        This method checks if the cache for the ID dictionary is empty. If it is,
+        it rebuilds the cache by calling the ``rebuild_cache`` method. Finally, it
+        returns the cached dictionary of segments.
+
+        Returns:
+            Dict[str, Segment]: A dictionary mapping segment IDs to Segment
+                objects.
+        """
+        if not self._id_dictionary_cache:
+            self.rebuild_cache()
+        return self._id_dictionary_cache
+
+    @property
+    def name_dictionary(self) -> Dict[str, Segment]:
+        """Return a dictionary of segments indexed by their names.
+
+        This method checks if the cache for the name dictionary is empty. If it is,
+        it rebuilds the cache by calling the ``rebuild_cache`` method. Finally, it
+        returns the cached dictionary of segments.
+
+        Returns:
+            Dict[str, Segment]: A dictionary where the keys are segment names and
+                the values are Segment objects.
+        """
+        if not self._name_dictionary_cache:
+            self.rebuild_cache()
+        return self._name_dictionary_cache
+
+    @property
+    def code_dictionary(self) -> Dict[str, Segment]:
+        """Return a dictionary of segments indexed by their codes.
+
+        This method checks if the cache for the code dictionary is empty. If it is,
+        it rebuilds the cache by calling the ``rebuild_cache`` method. Finally, it
+        returns the cached dictionary of segments.
+
+        Returns:
+            Dict[str, Segment]: A dictionary where the keys are segment codes and
+                the values are Segment objects.
+        """
+        if not self._code_dictionary_cache:
+            self.rebuild_cache()
+        return self._code_dictionary_cache
+
+    def to_list(self) -> List[Segment]:
+        """Return the list of segments.
+
+        Returns:
+            List[Segment]: The underlying list of segments.
+        """
+        return list(self)
