+++ PermutiveAPI/__init__.py
@@ -6,7 +6,9 @@ cohorts and workspaces through Permutive's REST API.  Refer to ``README.md`` in
 the repository root for installation and full usage documentation.
 """
 
-from PermutiveAPI.Source import Import, ImportList, Segment, SegmentList
+from PermutiveAPI.Source import Source
+from PermutiveAPI.Import import Import, ImportList
+from PermutiveAPI.Segment import Segment, SegmentList
 from PermutiveAPI.User import Identity, Alias
 from PermutiveAPI.Cohort import Cohort, CohortList
 from PermutiveAPI.Workspace import Workspace, WorkspaceList
@@ -19,6 +21,7 @@ __all__ = ["Cohort",
            "ImportList",
            "Segment",
            "SegmentList",
+           "Source",
            "Workspace",
            "WorkspaceList",
            "Identity",
